* Check for column before attempting to remove in migration. * Use closure for whereHas()

+++ database/migrations/2018_01_10_143606_add_active_field.php
@@ -27,7 +27,9 @@ class AddActiveField extends Migration
     public function down()
     {
         Schema::table('ctrts_schedules', function (Blueprint $table) {
-            $table->dropColumn('active');
+            if (Schema::hasColumn('ctrts_schedules', 'active')) {
+                $table->dropColumn('active');
+            }
         });
     }
 }
